Refactor Castable concern and move to domain-specific location
- Move HasCastableConcern from core/concerns to core/domains/cast/concerns
- Create new Castable service to handle casting logic
- Update BaseCastable, interfaces, and exceptions to new domain structure
- Refactor import paths across project to use new cast domain location
- Maintain existing casting functionality and type safety
