+++ src/core/domains/cast/base/BaseCastable.ts
@@ -1,7 +1,8 @@
-import HasCastableConcern from "@src/core/concerns/HasCastableConcern";
-import { IHasCastableConcern } from "@src/core/interfaces/concerns/IHasCastableConcern";
+import { IHasCastableConcern } from "@src/core/domains/cast/interfaces/IHasCastableConcern";
 import { ICtor } from "@src/core/interfaces/ICtor";
 import compose from "@src/core/util/compose";
+
+import HasCastableConcern from "../concerns/HasCastableConcern";
 
 const BaseCastable: ICtor<IHasCastableConcern> = compose(class {}, HasCastableConcern)
