+++ src/core/domains/cast/interfaces/CastException.ts
