+++ src/core/domains/cast/interfaces/IHasCastableConcern.ts
