+++ src/core/domains/events/base/BaseEvent.ts
@@ -1,4 +1,4 @@
-import BaseCastable from "@src/core/base/BaseCastable";
+import BaseCastable from "@src/core/domains/cast/base/BaseCastable";
 import { TCastableType, TCasts } from "@src/core/domains/cast/interfaces/IHasCastableConcern";
 import EventInvalidPayloadException from "@src/core/domains/events/exceptions/EventInvalidPayloadException";
 import { IBaseEvent } from "@src/core/domains/events/interfaces/IBaseEvent";
@@ -6,7 +6,6 @@ import IEventDriver from "@src/core/domains/events/interfaces/IEventDriver";
 import { IEventService } from "@src/core/domains/events/interfaces/IEventService";
 import { ICtor } from "@src/core/interfaces/ICtor";
 import { App } from "@src/core/services/App";
-
 abstract class BaseEvent<TPayload = unknown> extends BaseCastable implements IBaseEvent<TPayload> {
 
     protected payload: TPayload | null = null;
