+++ src/core/util/castObject.ts
@@ -1,5 +1,5 @@
 import BaseCastable from "@src/core/base/BaseCastable";
-import { TCasts } from "@src/core/interfaces/concerns/IHasCastableConcern";
+import { TCasts } from "@src/core/domains/cast/interfaces/IHasCastableConcern";
 
 const castObject = <ReturnType = unknown>(data: unknown, casts: TCasts): ReturnType => {
     return new BaseCastable().getCastFromObject<ReturnType>(data as Record<string, unknown>, casts)
